rec: move handlers of rec to handlers.rs
handlers in rec/mod.rs to rec/handlers.rs
